Styled the Snapshots History page a little and added some page info at the top so it kind of looks okay. It is nothing fancy but it gets the job done.
Started working on Markets Available page that shows the Markets available for a specific Snapshot.
This is going to require some condition checking both in html and ts class files...
If the hashId provided in the url='latest', we want to get the latest, otherwise we want to find an older/specific snapshot. If the url id isn't 'latest', then we need to make an api request using the given hashId in the url. However, that hashId may not exist or even be valid, so we need to account for the api sending us a 404 error or 400/500 etc errors. If it does those we need to show error info on the webpage. If the api comes back with status 200, then we need to map the json response to a correct object and display it properly in a table.