Played around with Observable a little. I still need to learn more about async and how to externalize it do a separate service that handles all my api requests.
For the moment all api http requests are made in the component that requires them, which I don't like but will hopefully fix soon.

I worked on the Markets Available component so it can get the JSON for a specific Snapshot id and map that data to a Snapshot object. If there was an error then I get the status code and custom error message given.
I also worked on the html template for this component so it shows expected page if request was successful and shows an error page if it failed with the status code and message to help explain to the user what is going on.

TODO:
Next up is to create a new component for the Markets Table that gets the Snapshot object from parent component and builds the table from that data.