The very basics of building a line chart are working. Data is displaying correctly but looks ugly, so there is a lot to work on. Not sure I want to use this chart package. I need to read a lot of documentation yet.
TODO: Fix X axis so it shows readable times and isn't so cluttered
TODO: Fix Y axis so it ranges from 0 min to 100 max. Makes it easier to read at-a-glance if we have that consistency and makes sense in terms of 'business logic' of this project.
TODO: Make line chart line/points smaller/thinner
TODO: Add functionality so it creates line data on the chart for each contract, not just one.