Implemented contract history component, it builds a specific contract's history based on market and contract id's. It shows some basic page info like market and contract names etc.
Contract history component also uses a newly created contract-history-table component that has all the history in table form. That component is finished as well.

TODO: Up next is creating the chart of this saved data. This is already done in the market-data component but on this page only this specific contract is charted. However, there are going to be analyzer data series that can be displayed like Simple Moving Average 10min/60min etc which can show trends and analysis.

I want to implement features I created for the Desktop version of Prediction Assistant so that it was is up now.