Fixed issue where movers: Mover[] was not updating in child movers-table component. I thought that of(movers).subscribe(...) would see an update occurring and call the next() function but it wasn't for some reason. I still don't understand why exactly. The movers array in the child was being updated because of @Input which I was confused about. I thought movers wasn't being updated but it was, the subscribe wasn't working the way I thought it did though.
Issue solved by implementing ngOnChanges so that when movers does update (which it was already doing), we update the table datasource etc, this change also occurs in the DOM which updates the View.

Also, added a link in movers-table to the market of that contract so we can easily get to the page showing everything happening. If one contract is down it might mean another is up and we want to easily get to the market to see what is going on there. not a tech fix, just a QoL/usability addition.